feat(vm): native dispatch for QuantHash coercions + builtins relocation (§1)
VM-native dispatch for `.Set`/`.Bag`/`.Mix`/`.SetHash`/`.BagHash` coercions over
list-like receivers, eliminating a §1 tree-walk fallback (advances interpreter-
execution-path removal toward the env/state fold). Mirrors the PR-5/PR-7
`try_native_*` pattern: a dedicated helper placed just before the catch-all (after
user-method resolution, so it never shadows a user `.Set`/etc method).

The fallback breadth was measured across 145 whitelist files; after `new`
(constructor, ③-blocked) the QuantHash coercions were the broadest pure-value
category (~11 files), so they are the natural next builtins 降ろし.

dedup + placement:
- new `src/builtins/quanthash_coerce.rs` holds the pure element-folding free
  functions `to_set`/`to_bag`/`to_mix` (+ helpers), moved out of `impl Interpreter`
  in `methods_collection.rs` (961 -> 315 lines). They only used `self` for mutual
  recursion, so they are genuinely pure.
- the interpreter's `dispatch_method_by_name_2` and `methods_dispatch_new`
  (mix_pair_weight) now delegate to the builtins functions = single authoritative
  implementation shared by VM and interpreter.
- `is_lazy_for_coerce`/`is_lazy_for_set_ops` stay in runtime (used by 6 other
  files) and are made `pub(crate)` so builtins can reference them.

Conservative fallthrough (returns None -> interpreter): `.MixHash` (registers
container type metadata = interpreter-owned state) and Instance/Package receivers
(`__baggy_data__`, type objects, user coercion). Behavior-invariant.

Measured: set-operation tests' Set/Bag/Mix fallback -> 0. pin
t/native-quanthash-coerce.t (26: dedup, weights, flatten rule, mutable variants,
set-op composition). cargo test 458/0, make test PASS, 29 whitelisted set/bag/mix
tests green. (set.t/bag.t are pre-existing non-whitelist BLOCKERS-documented
failures unrelated to this change.)
