+++ src/builtins/mod.rs
@@ -12,6 +12,7 @@ pub(crate) mod methods_0arg;
 mod methods_narg;
 pub(crate) mod parse_base;
 pub(crate) mod primality;
+pub(crate) mod quanthash_coerce;
 pub(crate) mod rng;
 pub(crate) mod split;
 pub(crate) mod str_increment;
